Remove the --no-optimizations option that was causing setup failures
This caused the setup script to fail on "yarn build" due to the unrecognised option.

Unsure at time of writing if this is an outdated option on the `strapi build` command that no longer works with
v4. However it doesn't appear to be a valid option for `strapi build` in strapi v3 either.